Move drink button to other file
